Feat: custom event and refactor
