+++ build.gradle
@@ -3,7 +3,7 @@ plugins {
 }
 
 group = 'com.doubledeltas'
-version = '1.3.0'
+version = '1.3.1'
 
 repositories {
     mavenCentral()
