+++ src/main/java/com/doubledeltas/minecollector/event/listener/PlayerLoginEventListener.java
@@ -1,4 +1,4 @@
-package com.doubledeltas.minecollector.event;
+package com.doubledeltas.minecollector.event.listener;
 
 import com.doubledeltas.minecollector.MineCollector;
 import com.doubledeltas.minecollector.data.DataManager;
